+++ example/lib/module/overlay_page.dart
@@ -9,7 +9,7 @@ class OverlayPage extends StatelessWidget {
   Widget build(BuildContext context) => ExtendedScaffold(
           appBar: AppBarText('Overlay'),
           isScroll: true,
-          padding: const EdgeInsets.all(20),
+          padding: const EdgeInsets.fromLTRB(20, 0, 20, 20),
           children: [
             ElevatedText('showSnackBar', onTap: () {
               showSnackBar(const SnackBar(content: BText('show SnackBar')));